+++ crates/pipeline-types/src/serde_with_context/serde_config.rs
@@ -68,7 +68,7 @@ impl Default for TimestampFormat {
     }
 }
 
-// Representation of the SQL `TIMESTAMP` type.
+// Representation of the SQL `DECIMAL` type.
 #[derive(Clone, Debug)]
 pub enum DecimalFormat {
     String,